wip: adding test cases for stream search
